+++ python/polars_online/_kwargs.py
@@ -103,6 +103,8 @@ class EwridgeKwargs(ExprKwargs, total=False):
     long_halflife: float | None
     solve_every: float | None
     max_rows_between_solves: int | None
+    window: float | None
+    window_every: int | None
 
 
 class RlsKwargs(ExprKwargs, total=False):